feat: #1740 separated hanlders, fix blurry
