+++ packages/elements/src/v2/src/components/image-cropper/handlers.ts
@@ -1,0 +1,89 @@
+// We resize the canvas down when saving on retina devices otherwise the image
+// will be double or triple the imageCropperPreview size.
+export const getResizedCanvas = (canvas, newWidth, newHeight) => {
+  const tmpCanvas = document.createElement('canvas')
+  tmpCanvas.width = newWidth
+  tmpCanvas.height = newHeight
+
+  const ctx = tmpCanvas.getContext('2d')
+  ctx.drawImage(canvas, 0, 0, canvas.width, canvas.height, 0, 0, newWidth, newHeight)
+
+  return tmpCanvas
+}
+
+export const generateDownload = (previewCanvas, crop) => {
+  if (!crop || !previewCanvas) {
+    return
+  }
+
+  const dpr = window.devicePixelRatio || 1
+  const canvas = dpr !== 1 ? getResizedCanvas(previewCanvas, crop.width, crop.height) : previewCanvas
+
+  canvas.toBlob(
+    blob => {
+      const previewUrl = window.URL.createObjectURL(blob)
+
+      const anchor = document.createElement('a')
+      anchor.download = 'cropPreview.png'
+      anchor.href = URL.createObjectURL(blob)
+      anchor.click()
+
+      window.URL.revokeObjectURL(previewUrl)
+    },
+    'image/png',
+    1,
+  )
+}
+
+export const drawCanvasAfterCrop = ({ completedCrop, previewCanvasRef, imgRef }) => () => {
+  if (!completedCrop || !previewCanvasRef.current || !imgRef.current) {
+    return
+  }
+
+  const image = imgRef.current
+  const canvas = previewCanvasRef.current
+  const crop = completedCrop
+  const dpr = window.devicePixelRatio || 1
+
+  const originWidth = image.naturalWidth
+  const originHeight = image.naturalHeight
+
+  const scaleX = originWidth / image.width
+  const scaleY = originHeight / image.height
+
+  const cropRatio = crop.width / crop.height
+
+  let outputWidth
+  let outputHeight
+
+  // get width or height, based on ratio
+  if (cropRatio > 1) {
+    outputWidth = originWidth
+    outputHeight = outputWidth / cropRatio
+  } else {
+    outputHeight = originHeight
+    outputWidth = outputHeight * cropRatio
+  }
+
+  const ctx = canvas.getContext('2d')
+
+  // set real dimension when download, different from display dimension
+  canvas.width = outputWidth
+  canvas.height = outputHeight
+
+  ctx.drawImage(
+    image,
+    crop.x * scaleX,
+    crop.y * scaleY,
+    crop.width * scaleX,
+    crop.height * scaleY,
+    0,
+    0,
+    outputWidth * dpr,
+    outputHeight * dpr,
+  )
+}
+
+export const onLoadHandle = imgRef => img => {
+  imgRef.current = img
+}
